feat(8): CSV product import

+++ src/main/java/com/slim/kreadevis_backend/dto/product/ProductRequest.java
@@ -3,12 +3,14 @@ package com.slim.kreadevis_backend.dto.product;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public record ProductRequest(
         @NotBlank String label,
         String description,
         @NotNull Long stockQuantity,
-        float unitPrice,
-        float vatRate,
+        BigDecimal unitPrice,
+        BigDecimal vatRate,
         String referenceCode,
         Long supplierId
 ) {}
