+++ src/main/java/com/slim/kreadevis_backend/dto/quote/QuoteItemResponse.java
@@ -2,10 +2,12 @@ package com.slim.kreadevis_backend.dto.quote;
 
 import com.slim.kreadevis_backend.dto.product.ProductResponse;
 
+import java.math.BigDecimal;
+
 public record QuoteItemResponse(
         Long id,
         ProductResponse product,
         Long quantity,
-        float unitPrice,
-        float totalPrice
+        BigDecimal unitPrice,
+        BigDecimal totalPrice
 ) {}
